requirements.txt
numpy
tqdm
torch
fla-core==0.4.2
huggingface-hub
kernels
setuptools
typing-extensions==4.15.0
datasets
tiktoken
sentencepiece
